+++ lib/scanner/apache.rb
@@ -39,6 +39,7 @@ module Yawast
         #run all the defined checks
         check_server_status(uri.copy)
         check_server_info(uri.copy)
+        check_tomcat_manager(uri.copy)
       end
 
       def self.check_server_status(uri)
@@ -47,6 +48,48 @@ module Yawast
 
       def self.check_server_info(uri)
         check_page_for_string uri, '/server-info', 'Apache Server Information'
+      end
+
+      def self.check_tomcat_manager(uri)
+        uri.path = '/manager/html'
+        uri.query = '' if uri.query != nil
+
+        ret = Yawast::Shared::Http.get(uri)
+
+        if ret.include? '<tt>conf/tomcat-users.xml</tt>'
+          #this will get Tomcat 7+
+          Yawast::Utilities.puts_warn "Apache Tomcat Manager page found: #{uri}"
+          check_tomcat_manager_passwords uri
+
+          puts ''
+        else
+          #check for Tomcat 6 and below
+          uri.path = '/manager'
+          ret = Yawast::Shared::Http.get(uri)
+
+          if ret.include? '<tt>conf/tomcat-users.xml</tt>'
+            Yawast::Utilities.puts_warn "Apache Tomcat Manager page found: #{uri}"
+            check_tomcat_manager_passwords uri
+
+            puts ''
+          end
+        end
+      end
+
+      def self.check_tomcat_manager_passwords(uri)
+        #check for known passwords
+        ret = Yawast::Shared::Http.get(uri, {'Authorization' => 'Basic dG9tY2F0OnRvbWNhdA=='}) #tomcat:tomcat
+        if ret.include? '<font size="+2">Tomcat Web Application Manager</font>'
+          Yawast::Utilities.puts_vuln 'Apache Tomcat Manager weak password: tomcat:tomcat'
+        end
+        ret = Yawast::Shared::Http.get(uri, {'Authorization' => 'Basic YWRtaW46YWRtaW4='}) #admin:admin
+        if ret.include? '<font size="+2">Tomcat Web Application Manager</font>'
+          Yawast::Utilities.puts_vuln 'Apache Tomcat Manager weak password: admin:admin'
+        end
+        ret = Yawast::Shared::Http.get(uri, {'Authorization' => 'Basic YWRtaW46'}) #admin:(blank)
+        if ret.include? '<font size="+2">Tomcat Web Application Manager</font>'
+          Yawast::Utilities.puts_vuln 'Apache Tomcat Manager weak password: admin:(blank)'
+        end
       end
 
       def self.check_page_for_string(uri, path, search)